Replace href-redirect with focus in <ValidationErrorSummaryBase />
Fixes a bug that caused the entire error summary to disappear when clicking one of the link elements,
which is a consqeuence of <FormikStep />'s history.listen-callback to reset formik's submitCount.

Note: the mentioned history.listen-callback is needed because we need a way of hooking into the browsers
behaviour when clicking the in-browser back button to navigate back, as we need to set formik's submitCount to 0.
This is done to ensure we don't still show the [outdated] error summary if user navigates back.