Implement <InformationPanel />
Implements InformationPanel component for rendering info text in <IntroPage /> component.
Renames everything previously named "Information*" such as buttons and icons to HelperText,
as it is more descriptive of what it does and would crash with the naming of the new
InformationPanel.

Replace custom styling in introPage for spacing with using <Box> component.