Rewrite file upload logic
FormikFileUploader/uploadAttachments was too hard to understand and it needed a rewrite.

In addition to rewriting it, this commit also ensures that user will always get some kind
of feedback when uploading attachments, specifically, the previous implementation did not
give user any feedback at all if the file that user attempted to upload, had an invalid extension.
This will now be shown as an error, in the exact same way as its handled when getting bad
responses after trying to upload a file.

Also moves the logic for deciding input[type=file]'s accept-attribute value out of <FileInputBase/>
and into <FormikFileInputProps>/<FileInput> and <FormikFileUploader> which will drill the prop value to
the base-component via props to FileInput. This might be changed back at a later stage, but
has been done this way for now, as I don't think this should be the responsibility of a
generic input-base kind of component to decide the accept-value for itself - it should probably be up to the consumer of
<FileInputBase> to make those decisions. Also, I found it to make more sense that this is
thightly connected to the rest of the validation logic etc. for attachments, which is primarily located
in FormikFileUploader.