Fix broken /soknad-call
Fixed a couple of errors during submission of Søknad due to a couple
of missing checks on HENT_BARN_FEATURE-toggle. Specifically, it failed
because of JSON.parse of barn fetched from api was done even when feature
was toggled off, and also, relasjon_til_barn field was being unset regardless
of feature-toggle which resulted in a 400 bad request.